Use identity operators (===) instead of equality (==) operators
https://stackoverflow.com/questions/359494/which-equals-operator-vs-should-be-used-in-javascript-comparisons

"JavaScript has two sets of equality operators: === and !==, and their
evil twins == and !=. The good ones work the way you would expect. If
the two operands are of the same type and have the same value, then ===
produces true and !== produces false. The evil twins do the right thing
when the operands are of the same type, but if they are of different
types, they attempt to coerce the values. the rules by which they do
that are complicated and unmemorable."